moved BasicRecurrentArchitecture to zoo
